+++ data-extraction/src/index.js
@@ -563,14 +563,15 @@ if (hasTodo('fishingLog')) {
           spot.TerritoryType = diademTerritory;
         }
         const c = spot.TerritoryType.Map.SizeFactor / 100.0;
+
         spots.push({
           id: spot.ID,
           mapId: spot.TerritoryType.Map.ID,
           placeId: spot.TerritoryType.PlaceName.ID,
           zoneId: spot.PlaceName.ID,
           coords: spot.ID >= 10000 ? diademFishingSpotCoords[spot.ID] : {
-            x: (41.0 / c) * (((spot.X * c)) / 2048.0) + 1,
-            y: (41.0 / c) * (((spot.Z * c)) / 2048.0) + 1
+            x: Math.floor(100 * (41.0 / c) * (spot.X / 2048.0) + 1) / 100,
+            y: Math.floor(100 * (41.0 / c) * (spot.Z / 2048.0) + 1) / 100
           },
           fishes: Object.keys(spot)
             .filter(key => /Item\dTargetID/.test(key))
@@ -588,7 +589,6 @@ if (hasTodo('fishingLog')) {
             if (fishes.indexOf(fish.ID) === -1) {
               fishes.push(fish.ID);
             }
-            const c = spot.TerritoryType.Map.SizeFactor / 100.0;
             const entry = {
               itemId: fish.ID,
               level: spot.GatheringLevel,
@@ -599,8 +599,8 @@ if (hasTodo('fishingLog')) {
               spot: {
                 id: spot.ID,
                 coords: {
-                  x: (41.0 / c) * ((spot.X * c) / 2048.0) + 1,
-                  y: (41.0 / c) * ((spot.Z * c) / 2048.0) + 1
+                  x: Math.floor(100 * (41.0 / c) * (spot.X / 2048.0) + 1) / 100,
+                  y: Math.floor(100 * (41.0 / c) * (spot.Z / 2048.0) + 1) / 100
                 }
               }
             };
@@ -1221,6 +1221,31 @@ if (hasTodo('jobCategories')) {
   }, null, () => {
     persistToTypescript('job-categories', 'jobCategories', jobCategories);
     done('jobCategories');
+  });
+}
+
+if (hasTodo('jobs')) {
+  const jobAbbrs = {};
+  const jobNames = {};
+  getAllPages('https://xivapi.com/ClassJob?columns=ID,Abbreviation_*,Name_*').subscribe(page => {
+    page.Results.forEach(job => {
+      jobNames[job.ID] = {
+        en: job.Name_en,
+        ja: job.Name_ja,
+        de: job.Name_de,
+        fr: job.Name_fr
+      };
+      jobAbbrs[job.ID] = {
+        en: job.Abbreviation_en,
+        ja: job.Abbreviation_ja,
+        de: job.Abbreviation_de,
+        fr: job.Abbreviation_fr
+      };
+    });
+  }, null, () => {
+    persistToJsonAsset('job-name', jobNames);
+    persistToJsonAsset('job-abbr', jobAbbrs);
+    done('jobs');
   });
 }